+++ protected/humhub/docs/CHANGELOG.md
@@ -19,7 +19,9 @@ HumHub Change Log
 - Enh: Added `humhub\modules\file\models\isAssignedTo($record)` argument in order to check if a File is attached to a specific record
 - Fix: rich-text mobile view wrong min-height calculation
 - Fix #3314: layout container width differences
-
+- Fix comment scroll overflows button
+- Fix widgetAction events case issue
+- Added `humhub.modules.util.string.capitalizeFirstLetter` and `lowerCaseFirstLetter`
 
 1.3.4  (September 25, 2018)
 ---------------------------